pyproject.toml
[tool.poetry]
name = "taxdocs"
version = "0.1.0"
description = ""
authors = ["Michael Bianco <iloveitaly@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
# funcy-pipe = "^0.11.0"
funcy-pipe = { path = "/Users/mike/Projects/python/funcy-pipe", develop = true }
click = "^8.1.7"
griptape = { path = "/Users/mike/Projects/ai/griptape", develop = true }
pandas = "^2.2.1"
xlrd = "^2.0.1"
boto3 = "^1.34.80"
openpyxl = "^3.1.2"
xlutils = "^2.0.0"
pillow = "^10.3.0"
pdf2image = "^1.17.0"
cachetools = "^5.3.3"
pytesseract = "^0.3.10"
whatever = "^0.7"
structlog = "^24.1.0"
python-decouple-typed = "^3.11.0"
schema = "^0.7.5"
filetype = "^1.2.0"
jinja2 = "^3.1.3"
google-cloud-vision = "^3.7.2"
azure-ai-documentintelligence = "^1.0.0b3"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
pyright = "^1.1.357"
pytest = "^8.1.1"
pytest-check = "^2.3.1"
pretty-traceback = "^2023.1020"
autoflake = "^2.3.1"
autoimport = "^1.4.0"
deptry = "^0.16.1"
flake8 = "^7.0.0"
flake8-builtins = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# [tool.black]
# line-length = 120

[tool.isort]
profile = "black"
# helpful when copy/pasting GPT code
float_to_top = true

[tool.pytest.ini_options]
# addopts = "--pdbcls=IPython.terminal.debugger:Pdb"
addopts = "--pdbcls=pdbr:RichPdb"
